+++ point.cc
@@ -7,7 +7,7 @@ Point::Point() {
     z = 0.0;
 }
 
-Point::Point(float xn, float yn, float zn) {
+Point::Point(double xn, double yn, double zn) {
     x = xn;
     y = yn;
     z = zn;
@@ -21,15 +21,15 @@ Point Point::operator+ (const Vec3 &v) const {
     return Point(x + v.x, y + v.y, z + v.z);
 }
 
-void Point::setX(float xn) {
+void Point::setX(double xn) {
     x = xn;
 }
 
-void Point::setY(float yn) {
+void Point::setY(double yn) {
     y = yn;
 }
 
-void Point::setZ(float zn) {
+void Point::setZ(double zn) {
     z = zn;
 }
