+++ point.h
@@ -5,14 +5,14 @@
 
 class Point {
 protected:
-    float x, y, z;
+    double x, y, z;
 
 public:
     Point();
-    Point(float, float, float);
-    void setX(float);
-    void setY(float);
-    void setZ(float) ;
+    Point(double, double, double);
+    void setX(double);
+    void setY(double);
+    void setZ(double) ;
     Vec3 operator- (const Point&) const;
     Point operator+ (const Vec3 &v) const;
